+++ client/components/Card.js
@@ -1,43 +1,7 @@
 import { ImNewTab } from "react-icons/im";
-import { FaRegHeart } from "react-icons/fa";
-import { FaHeart } from "react-icons/fa";
-import React, { useState, useEffect } from "react";
-import { auth, firestore } from '../firebase/firebaseConfig';
-import { doc, getDoc, updateDoc } from 'firebase/firestore';
+import Favorite from "./Favorite";
 
 const Card = ({data, showImg=false, showHeart=false}) => {
-    const [favorite, setFavorite] = useState(false);
-    const [signedIn, setSignedIn] = useState(false);
-
-    useEffect(() => {
-        checkFavoriteDatabase();
-        checkForUserSignedIn();
-    },[data]);
-    const checkForUserSignedIn = async () => {
-        const currentUser = auth.currentUser;
-        if(currentUser) {
-            setSignedIn(true);
-        }
-    }
-    const checkFavoriteDatabase = async () => {
-        const currentUser = auth.currentUser;
-        if (currentUser) {
-            const userId = currentUser.uid;
-            const userRef = doc(firestore, "users", userId);
-            const userDoc = await getDoc(userRef);
-            if (userDoc.exists()) {
-                const {favoriteSchools} = userDoc.data();
-                if(favoriteSchools) {
-                    const found = favoriteSchools.includes(data?.dbn);
-                    setFavorite(found);
-                }
-                else{
-                    const found = false;
-                    setFavorite(found);
-                }
-            }
-        }
-    }
 
     if(!data?.address) {
         data = { ...data, address: "Unavailable"};
@@ -53,43 +17,7 @@ const Card = ({data, showImg=false, showHeart=false}) => {
           data = { ...data, website: "https://"+data.website};
         }
     }
-    const handleFavoriteToggle = async () => {
-        if (favorite) {
-            const currentUser = auth.currentUser;
-        if (currentUser) {
-            const userId = currentUser.uid;
-            const userRef = doc(firestore, "users", userId);
-            const userDoc = await getDoc(userRef);
-            if (userDoc.exists()) {
-                const {favoriteSchools} = userDoc.data();
-                if(favoriteSchools) {
-                    const updateFavoriteSchools = favoriteSchools.filter(schoolDbn=>schoolDbn!==data?.dbn);
-                    await updateDoc(userRef, {favoriteSchools: updateFavoriteSchools});
-                 }
-            }
-        }
-            setFavorite(false);
-        } else {
-            const currentUser = auth.currentUser;
-            if (currentUser) {
-                const userId = currentUser.uid;
-                const userRef = doc(firestore, "users", userId);
-                const userDoc = await getDoc(userRef);
-                if (userDoc.exists()) {
-                    const userData = userDoc.data();
-                    if(!userData.favoriteSchools) {
-                        await updateDoc(userRef, {favoriteSchools:[data?.dbn]});
-                    }
-                    else {
-                        await updateDoc(userRef, {favoriteSchools: [...userData.favoriteSchools, data?.dbn]});
-                    }
-                }
-            }
-            setFavorite(true);
-        }
-    };
    
-
     return (
         <div className="card-fill">
             <div className="card-body">
@@ -133,11 +61,11 @@ const Card = ({data, showImg=false, showHeart=false}) => {
                         </h6>
                     </div>
                     {data?.imgUrl && showImg && <img src={data.imgUrl} style={{ width: "240px", height: "190px",marginLeft: "20px", objectFit: "contain", marginRight:"20px"}} />}
-                    {signedIn && showHeart &&
-                        <button className="favorite-button" onClick={handleFavoriteToggle}>
-                            {favorite ?  <FaHeart style={{color:"red"}} /> :  <FaRegHeart />}
-                        </button>
-                }
+                    <div className="fav-map">
+                        {showHeart &&
+                            <Favorite data={data}/>
+                        }
+                    </div>
                 </div>
             </div>
         </div>
