+++ client/pages/schools/quality-reports/[dbn].js
@@ -2,10 +2,7 @@ import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import CardSquare from "../../../components/CardSquare";
 import SchoolButton from "../../../components/SchoolButton";
-import { FaRegHeart } from "react-icons/fa";
-import { FaHeart } from "react-icons/fa";
-import { auth, firestore } from '../../../firebase/firebaseConfig';
-import { doc, getDoc, updateDoc } from 'firebase/firestore';
+import Favorite from "@/components/Favorite";
 
 
 function SchoolQualityReport() {
@@ -13,10 +10,7 @@ function SchoolQualityReport() {
     const { dbn } = router.query;
     const [school, setSchool] = useState([]);
     const[message, setMessage] = useState("Loading");
-    const [favorite, setFavorite] = useState(false);
-    const [signedIn, setSignedIn] = useState(false);
     
-
     useEffect(() => {
         if (dbn) {
             fetch(process.env.NEXT_PUBLIC_SERVER_URL + `/api/schools/quality-reports/${dbn}`)
@@ -32,82 +26,14 @@ function SchoolQualityReport() {
         }
     }, [dbn]);
 
-    useEffect(() => {
-        checkForUserSignedIn();
-        checkFavoriteDatabase();
-      },[school.dbn]);
-
-      const checkForUserSignedIn = async () => {
-        const currentUser = auth.currentUser;
-        if(currentUser) {
-            setSignedIn(true);
-        }
-      }
-      const checkFavoriteDatabase = async () => {
-          const currentUser = auth.currentUser;
-          if (currentUser) {
-              const userId = currentUser.uid;
-              const userRef = doc(firestore, "users", userId);
-              const userDoc = await getDoc(userRef);
-              if (userDoc.exists()) {
-                  const {favoriteSchools} = userDoc.data();
-                  if(favoriteSchools) {
-                      const found = favoriteSchools.includes(school.dbn);
-                      setFavorite(found);
-                  }
-                  else{
-                      setFavorite(false);
-                  }
-              }
-          }
-      }
-  
-      const handleFavoriteToggle = async () => {
-        if (favorite) {
-            const currentUser = auth.currentUser;
-        if (currentUser) {
-            const userId = currentUser.uid;
-            const userRef = doc(firestore, "users", userId);
-            const userDoc = await getDoc(userRef);
-            if (userDoc.exists()) {
-                const {favoriteSchools} = userDoc.data();
-                if(favoriteSchools) {
-                    const updateFavoriteSchools = favoriteSchools.filter(schoolDbn=>schoolDbn!==school.dbn);
-                    await updateDoc(userRef, {favoriteSchools: updateFavoriteSchools});
-                 }
-            }
-        }
-            setFavorite(false);
-        } else {
-            const currentUser = auth.currentUser;
-            if (currentUser) {
-                const userId = currentUser.uid;
-                const userRef = doc(firestore, "users", userId);
-                const userDoc = await getDoc(userRef);
-                if (userDoc.exists()) {
-                    const userData = userDoc.data();
-                    if(!userData.favoriteSchools) {
-                        await updateDoc(userRef, {favoriteSchools:[school.dbn]});
-                    }
-                    else {
-                        await updateDoc(userRef, {favoriteSchools: [...userData.favoriteSchools, school.dbn]});
-                    }
-                }
-            }
-            setFavorite(true);
-        }
-    };
-
     if (message!="School data not found!") {
         return (
             <div className="background-color">
             <section id="quality-reports">
                 <div className="container">
-                {signedIn && 
-                        <button className="favorite-button-school-page" onClick={()=>handleFavoriteToggle()}>
-                            {favorite ?  <FaHeart style={{color:"red"}} /> :  <FaRegHeart />}
-                        </button>
-                }
+                    <div className="favorite-button-school-page">
+                        <Favorite data={school}/>
+                    </div>
                     <div className="row">
                         <h1 className="display-1">{school.school_name}</h1>
                         <p className="desc">What is a School Quality Report?</p>
