+++ client/styles/globals.css
@@ -431,21 +431,19 @@ label #cancel{
 }
 
 .favorite-button {
-  position: absolute;
-  bottom: 15px; 
-  right: 15px; 
   background: none;
   border: none;
   cursor: pointer;
   font-size: 24px;
 }
+.fav-map {
+  bottom: 15px; 
+  right: 15px; 
+  position: absolute;
+}
 .favorite-button-school-page {
   position: absolute;
-  /* bottom: 15px;  */
   right: 15px; 
-  background: none;
-  border: none;
-  cursor: pointer;
   font-size: 30px;
 }
 .delete-button {
